fix: Update Header component and correct typos
The Header component was updated to include new enhancements for user interactivity. This includes the addition of a Link, Logo, and an AvatarDropdown to implement conditional rendering for sign in and user information. Improvements in the props used and adjustments in the styling also took place for better structuring.

In addition, corrected multiple instances of 'AvtarDropdown' to 'AvatarDropdown' to standardize component naming conventions across the project. This change also required modifications in file names and usage instances in the storybook.