Refactor auth checks and handling for all components
This commit streamlines the handling of user authentication across all the components of the project. All unnecessary or redundant checks for user authentication have been removed from components. User data fetching logic has been consolidated into the `ProtectedRoutes` and `Layout` components, which handle user redirection to the 'sign-in' page if unauthenticated or if an error occurs. This refactor removes the need for additional navigation and error-checking related to authentication within each component, improving efficiency and cohesiveness of code. Additionally, the redundant 'null' type has been removed from the 'UserType' property in 'header.tsx', and a console log removed, further tidying the codebase.