"Update navigations and HTTP requests Redirecting"
Update the href in the recoveryEmail from 'confirm-email' to 'reset-password' to match the actual intended functionality. On successful sign-in, users are now redirected to the index '/' page instead of '/packs'. Also, upon successful password reset, users are now navigated back to the index page. Remove loading states from the layout file to streamline the UX. Adjust the service base URL in 'auth.api.ts' and include token in request headers to match with backend changes. On sign-out, the 'Me' tag is invalidated to prompt data refetch on subsequent sign-ins. Renamed exported api from 'authAPI' to 'api' for better intuitiveness. Store adjustments includes removing comments and unused code, replacing 'authAPI' with 'authReducer'. Put back 'authAPI' inside 'auth' object of the reducer. Lastly, added a new slice to manage user and token in Redux state. This commit aims to provide a smoother user experience and adapt to changes on the backend.