Remove unused code and file
Deleted the "icons.stories.tsx" file and removed an unused variable from "layout.tsx". The "icons.stories.tsx" was not actually used in the application, hence it was removed. Similarly, the variable "isLoading" was declared but not used in "layout.tsx", therefore it was removed to clean up the code.